Initialize Base Code State
NOTE: The software in its current state has been tested and confirmed operational (separately as a worker for the Discord bot and a web dyno for the Flask webapp).

WARNING: In order to operate this software, certain files are required which are not present in the GitHub repository for security reasons.
Future changes to README.md will explain exactly how to achieve this.
In short, the changes required are as follows:
1. Both apps need a "bucket.txt" file
2. The discord bot app needs a "token.txt" file
3. Both apps need their own unique versions of the following files:
a. Procfile
b. requirements.txt
c. .gitignore

This commit represents the initial code state for development of the "nextlevel" project.

Recently I have privately developed two projects:
1. A discord bot with persistent AWS S3 storage
2. A flask webapp with persistent AWS S3 storage

This commit serves to publicize these previously-private files, and simultaneously to integrate them into a single software package with as little redundant code as possible.